Revert "Add default sort option"
This reverts commit 6b7b6fd246117649009d73c569e8f089ec4e01bb.
