+++ CHANGELOG.md
@@ -11,7 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add command (`whichkey.undoKey`) to undo entered key.
 - Add menu buttons to which key menu.
-- Add a default sorting order for bindings.
 
 ### Changed
