+++ src/charCode.ts
@@ -24,11 +24,6 @@ export const enum CharCode {
     Z = 90,
 
     /**
-     * The `[` character.
-     */
-    LeftBracket = 91,
-
-    /**
      * The `\` character.
      */
     Backslash = 92,
