+++ src/constants.ts
@@ -21,7 +21,6 @@ export enum CommandKey {
 }
 
 export enum SortOrder {
-	Default = 'default',
 	None = 'none',
 	Alphabetically = 'alphabetically',
 	NonNumberFirst = 'nonNumberFirst',
